split wheel tests
